docs(api): first pass
