+++ .npmignore
@@ -32,3 +32,4 @@ build
 node_modules
 
 test
+docs
