+++ .travis.yml
@@ -1,11 +1,17 @@
 sudo: false
 language: node_js
-node_js:
-  - 4
-  - 5
-  - stable
 
-# Make sure we have new NPM.
+matrix:
+  include:
+  - node_js: 4
+    env: CXX=g++-4.8
+  - node_js: 6
+    env:
+      - SAUCE=true
+      - CXX=g++-4.8
+  - node_js: stable
+    env: CXX=g++-4.8
+
 before_install:
   - npm install -g npm
 
@@ -22,4 +28,9 @@ after_success:
   - npm run coverage-publish
 
 addons:
-  firefox: 'latest'
+  firefox: latest
+  apt:
+    sources:
+      - ubuntu-toolchain-r-test
+    packages:
+      - g++-4.8
