+++ package.json
@@ -2,8 +2,7 @@
   "name": "multibase",
   "version": "0.2.0",
   "description": "JavaScript implementation of the multibase specification",
-  "main": "lib/index.js",
-  "jsnext:main": "src/index.js",
+  "main": "src/index.js",
   "scripts": {
     "test:node": "aegir-test node",
     "lint": "aegir-lint",
@@ -33,7 +32,7 @@
     "formats"
   ],
   "devDependencies": {
-    "aegir": "^6.0.1",
+    "aegir": "^9.1.2",
     "chai": "^3.5.0",
     "pre-commit": "^1.1.3"
   },
